apps/bot/pyproject.toml
[tool.poetry]
name = "travel-planner"
version = "0.1.0"
description = "Telegram bot for planning your travels"
authors = ["Herbstblatt <48344463+Herbstblatt@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
aiogram-dialog = "^2.1.0"
asyncpg = "^0.29.0"
fluent-runtime = "^0.4.0"
aerich = "^0.7.2"
tortoise-orm = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[tool.aerich]
tortoise_orm = "main.tortoise_config"
location = "./db/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
